start-server.bat
I've added the information about using start-server.bat in a background process for local testing to the DEVELOPMENT RULES section. This will help prevent issues where the server stops running when trying to test in the same session, as the batch file is blocking. The instructions now clearly specify that the server should be started in a background process for proper testing workflow.